+++ app-modules/division/src/Filament/Resources/Divisions/Pages/ViewDivision.php
@@ -34,9 +34,9 @@
 </COPYRIGHT>
 */
 
-namespace AdvisingApp\Division\Filament\Resources\DivisionResource\Pages;
+namespace AdvisingApp\Division\Filament\Resources\Divisions\Pages;
 
-use AdvisingApp\Division\Filament\Resources\DivisionResource;
+use AdvisingApp\Division\Filament\Resources\Divisions\DivisionResource;
 use AdvisingApp\Division\Models\Division;
 use App\Filament\Resources\NotificationSettingResource;
 use App\Filament\Resources\UserResource;
